osd: Improve SnapMapper handling of inconsistencies
My using nop() instead of setattr() we avoid crashing
a replica that is missing the object for some reason.
Although this is a corruption, it would be getting "fix"
by removal of the snapshot.

In the other cases we accept completely missing keys (ENOENT)
that happens when an object is removed somehow or missed
by recovery/backfill.

Signed-off-by: David Zafman <dzafman@redhat.com>