cephfs-journal-tool: dump to per rank output file wherever necessary
cephfs-journal-tool supports operations on all ranks. Operations
such as dump/export do not write to distinct filenames hence
overwriting the data dumped or exported for the previous rank.

With this change (and further commits), for operations on all
ranks, dump/export would write to distinct filenames suffixed
by the mds rank (.0, .1, etc..). For operations on a single rank
or if there exist a single rank, the passed in filename is used
as it is.

Signed-off-by: Venky Shankar <vshankar@redhat.com>
(cherry picked from commit cb19db75c7f23278d2598b58526d1bf8bf00b679)

 Conflicts:
	src/tools/cephfs/JournalTool.h