Add several types to ceph-dencoder.
Add types: RGWBucketEntryPoint obj_version rgw_user
These are structures that are visible as data at rest
inside of rados when a bucket is made via radosgw.

RGWBucketEntryPoint is the contents of a rados
object with names that may be either "<bucket-name>"
or "<tenant>/<bucket-name>"

rgw_user is a structure contained inside of RGWBucketEntryPoint
and other structures.

obj_version is visible as the xattr "ceph.objclass.version"
on rados objects in ".rgw.meta" that contain ".bucket.meta."

Signed-off-by: Marcus Watts <mwatts@redhat.com>