mgr/ansible: Ansible orchestrator module
A Ceph Manager Orchestrator that uses a external REST API service to execute Ansible playbooks.

Modifications/New things introduced:

- Multiple inheritance removed for operations. Simplified classes dependencies
- Module Configuration variables introduced:[server_addr, server_port, username, password, certificate]
- Ansible runner service Client created. Manage all the interaction with the External Manager Service
- PlaybookExecution Class created. Manage all the life cycle of one playbook execution
- Modified Serve method. This is not an standby module. (not needed to loop forever)
- Modified wait method. Now this method is the engine that make ansible operations to progress and to obtain RESULTS
- Module adapted to work with Orchestator-cli module

Bug found and solved:
The OrchestratorCli_wait method did not work properly.
The main loop in this method must be executed if there is any completion object not completed.
It has been fixed to do that.

Signed-off-by: Juan Miguel Olmo Martínez <jolmomar@redhat.com>